Output XFDF fields into XML
Write `<field>` elements to the XFDF XML based on the cons cells in the
`fields` input list.

Change true and nil values to the FDF/XFDF checkbox value format of
"Yes" and "Off".

Also decided to use lowercase `t` to match the case of `nil`. Not sure
if that makes sense.